Call a time intensive function from within the Tkinter GUI as a test for the future. Remove the test list that is displayed upon power up.
