+++ bin/evt_viewer.py
@@ -1,11 +1,9 @@
 #! python
 from winlogtimeline import ui
 
-
 def main():
     app = ui.GUI()
     app.mainloop()
-
 
 if __name__ == "__main__":
     main()
